Reorganize the UrlOperations implementation in a dedicated module
`datalad_next.url_operations`
